file/directory/helper API reorg
