+++ src/IdentityServer/Services/Default/DefaultSessionManagementService.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Duende Software. All rights reserved.
 // See LICENSE in the project root for license information.
 
-using Duende.IdentityServer.Hosting.TicketStore;
 using Duende.IdentityServer.Models;
 using Duende.IdentityServer.Stores;
 using System.Collections.Generic;
