+++ src/Storage/Extensions/PersistedGrantFilterExtensions.cs
@@ -20,10 +20,10 @@ public static class PersistedGrantFilterExtensions
     {
         if (filter == null) throw new ArgumentNullException(nameof(filter));
 
-        if (String.IsNullOrWhiteSpace(filter.ClientId) &&
+        if ((String.IsNullOrWhiteSpace(filter.ClientId) || filter.ClientIds == null)&&
             String.IsNullOrWhiteSpace(filter.SessionId) &&
             String.IsNullOrWhiteSpace(filter.SubjectId) &&
-            String.IsNullOrWhiteSpace(filter.Type))
+            (String.IsNullOrWhiteSpace(filter.Type) || filter.Types == null))
         {
             throw new ArgumentException("No filter values set.", nameof(filter));
         }
