+++ src/Storage/Stores/IServerSideSessionStore.cs
@@ -2,7 +2,6 @@
 // See LICENSE in the project root for license information.
 
 using Duende.IdentityServer.Models;
-using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -51,62 +50,4 @@ public interface IServerSideSessionStore
     /// Queries sessions based on filter
     /// </summary>
     Task<QueryResult<ServerSideSession>> QuerySessionsAsync(SessionQuery filter = null, CancellationToken cancellationToken = default);
-}
-
-/// <summary>
-/// Filter to query user sessions
-/// </summary>
-public class SessionFilter
-{
-    /// <summary>
-    /// The subject ID
-    /// </summary>
-    public string SubjectId { get; init; }
-
-    /// <summary>
-    /// The sesion ID
-    /// </summary>
-    public string SessionId { get; init; }
-
-    /// <summary>
-    /// Validates
-    /// </summary>
-    public void Validate()
-    {
-        if (String.IsNullOrWhiteSpace(SubjectId) && String.IsNullOrWhiteSpace(SessionId))
-        {
-            throw new ArgumentNullException("SubjectId or SessionId is required.");
-        }
-    }
-}
-
-/// <summary>
-/// Filter to query all user sessions
-/// </summary>
-public class SessionQuery
-{
-    /// <summary>
-    /// The page number
-    /// </summary>
-    public int Page { get; set; }
-
-    /// <summary>
-    /// The number to return
-    /// </summary>
-    public int Count { get; set; }
-
-    /// <summary>
-    /// The subject ID
-    /// </summary>
-    public string SubjectId { get; init; }
-
-    /// <summary>
-    /// The sesion ID
-    /// </summary>
-    public string SessionId { get; init; }
-
-    /// <summary>
-    /// The user display name
-    /// </summary>
-    public string DisplayName { get; init; }
 }
