+++ src/Storage/Stores/Serialization/ClaimsPrincipalLite.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Duende Software. All rights reserved.
+// Copyright (c) Duende Software. All rights reserved.
 // See LICENSE in the project root for license information.
 
 #pragma warning disable 1591
